@@ -10,25 +10,26 @@
   <url>http://maven.apache.org</url>
 
   <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <jdk.version>1.8</jdk.version>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
+    <jdk.version>1.11</jdk.version>
     <oms.version>1.0-SNAPSHOT</oms.version>
     <servletapi.version>4.0.1</servletapi.version>
-    <spring.jdbc.version>5.1.8.RELEASE</spring.jdbc.version>
+    <spring.jdbc.version>5.3.14</spring.jdbc.version>
     <postgres.version>42.2.2</postgres.version>
     <dbcp2.version>2.6.0</dbcp2.version>
     <disruptor.version>3.4.2</disruptor.version>
+    <metrics.core.version>4.2.7</metrics.core.version>
     <junit.version>3.8.1</junit.version>
     <jbrcypt.version>0.4</jbrcypt.version>
-    <spring.ldap.version>2.3.1.RELEASE</spring.ldap.version>
+    <spring.ldap.version>2.3.5.RELEASE</spring.ldap.version>
     <start-class>com.ntw.auth.WebApplication</start-class>
   </properties>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.13.RELEASE</version>
+    <version>2.6.2</version>
   </parent>
 
   <dependencies>
@@ -72,7 +73,7 @@
     <dependency>
       <groupId>io.dropwizard.metrics</groupId>
       <artifactId>metrics-core</artifactId>
-      <version>3.2.2</version>
+      <version>${metrics.core.version}</version>
     </dependency>
 
     <dependency>
@@ -103,6 +104,12 @@
     </dependency>
 
     <dependency>
+      <groupId>javax.inject</groupId>
+      <artifactId>javax.inject</artifactId>
+      <version>1</version>
+    </dependency>
+
+    <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>${junit.version}</version>
@@ -128,7 +135,7 @@
     <plugins>
       <plugin>
         <artifactId>maven-war-plugin</artifactId>
-        <version>3.0.0</version>
+        <version>3.3.2</version>
         <configuration>
           <attachClasses>true</attachClasses>
           <classesClassifier>classes</classesClassifier>
