@@ -9,25 +9,25 @@
   <url>http://maven.apache.org</url>
 
   <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <jdk.version>1.8</jdk.version>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
+    <jdk.version>1.11</jdk.version>
     <servletapi.version>4.0.1</servletapi.version>
-    <spring.version>5.2.12.RELEASE</spring.version>
-    <slf4j.version>1.7.26</slf4j.version>
-    <log4j.version>2.12.0</log4j.version>
-    <jjwt.version>0.10.5</jjwt.version>
-    <jjwt.jackson.version>2.9.8</jjwt.jackson.version>
+    <spring.version>5.3.14</spring.version>
+    <spring.security.version>5.6.1</spring.security.version>
+    <eureka.client.version>3.1.0</eureka.client.version>
+    <spring.opentracing>0.5.9</spring.opentracing>
+    <jjwt.version>0.11.2</jjwt.version>
+    <jjwt.jackson.version>2.11.4</jjwt.jackson.version>
     <jaegar.version>3.3.1</jaegar.version>
-    <spring.boot.actuator.version>2.5.2</spring.boot.actuator.version>
-    <micrometer.version>1.7.1</micrometer.version>
+    <micrometer.version>1.8.1</micrometer.version>
     <junit.version>3.8.1</junit.version>
   </properties>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.13.RELEASE</version>
+    <version>2.6.2</version>
   </parent>
 
   <dependencies>
@@ -40,9 +40,17 @@
     </dependency>
 
     <dependency>
+      <groupId>io.opentracing.contrib</groupId>
+      <artifactId>opentracing-spring-cloud-starter</artifactId>
+      <version>${spring.opentracing}</version>
+    </dependency>
+
+
+
+    <dependency>
       <groupId>org.springframework.security</groupId>
       <artifactId>spring-security-web</artifactId>
-      <version>5.1.5.RELEASE</version>
+      <version>${spring.security.version}</version>
     </dependency>
     <dependency>
       <groupId>org.springframework</groupId>
@@ -81,7 +89,7 @@
     <dependency>
       <groupId>org.springframework.cloud</groupId>
       <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
-      <version>2.2.9.RELEASE</version>
+      <version>${eureka.client.version}</version>
       <exclusions>
         <exclusion>
           <groupId>org.springframework.security</groupId>
@@ -95,6 +103,10 @@
           <groupId>xpp3</groupId>
           <artifactId>xpp3_min</artifactId>
         </exclusion>
+        <exclusion>
+          <groupId>javax.servlet</groupId>
+          <artifactId>servlet-api</artifactId>
+        </exclusion>
       </exclusions>
     </dependency>
 
@@ -102,11 +114,16 @@
       <groupId>io.opentracing.contrib</groupId>
       <artifactId>opentracing-spring-jaeger-cloud-starter</artifactId>
       <version>${jaegar.version}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>io.opentracing.contrib</groupId>
+          <artifactId>opentracing-spring-cloud-starter</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-actuator</artifactId>
-      <version>${spring.boot.actuator.version}</version>
     </dependency>
     <dependency>
       <groupId>io.micrometer</groupId>
