@@ -10,20 +10,20 @@
   <url>http://maven.apache.org</url>
 
   <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
+    <jdk.version>1.11</jdk.version>
     <oms.version>1.0-SNAPSHOT</oms.version>
     <servletapi.version>4.0.1</servletapi.version>
-    <spring.jdbc.version>5.1.8.RELEASE</spring.jdbc.version>
-    <spring.cassandra.version>2.1.9.RELEASE</spring.cassandra.version>
+    <spring.jdbc.version>5.3.14</spring.jdbc.version>
     <postgres.version>42.2.2</postgres.version>
     <rabbitmq.version>5.12.0</rabbitmq.version>
-    <dbcp2.version>2.6.0</dbcp2.version>
     <okhttp.version>4.9.1</okhttp.version>
-    <slf4j.version>1.7.26</slf4j.version>
-    <log4j.version>2.12.0</log4j.version>
-    <disruptor.version>3.4.2</disruptor.version>
     <jaeger.client.version>1.3.2</jaeger.client.version>
+    <ribbon.client.version>2.2.10.RELEASE</ribbon.client.version>
+    <dbcp2.version>2.9.0</dbcp2.version>
+    <disruptor.version>3.4.4</disruptor.version>
+    <metrics.core.version>4.2.7</metrics.core.version>
     <junit.version>3.8.1</junit.version>
     <start-class>com.ntw.oms.order.WebApplication</start-class>
   </properties>
@@ -31,7 +31,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.13.RELEASE</version>
+    <version>2.6.2</version>
   </parent>
 
   <dependencies>
@@ -80,7 +80,7 @@
     <dependency>
       <groupId>io.dropwizard.metrics</groupId>
       <artifactId>metrics-core</artifactId>
-      <version>3.2.2</version>
+      <version>${metrics.core.version}</version>
     </dependency>
 
     <dependency>
@@ -120,7 +120,7 @@
     <dependency>
       <groupId>org.springframework.cloud</groupId>
       <artifactId>spring-cloud-starter-netflix-ribbon</artifactId>
-      <version>2.2.9.RELEASE</version>
+      <version>${ribbon.client.version}</version>
       <exclusions>
         <exclusion>
           <groupId>org.springframework.security</groupId>
@@ -130,13 +130,29 @@
           <groupId>com.google.code.findbugs</groupId>
           <artifactId>annotations</artifactId>
         </exclusion>
+        <exclusion>
+          <groupId>org.springframework.cloud</groupId>
+          <artifactId>spring-cloud-starter</artifactId>
+        </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>org.springframework.cloud</groupId>
+      <artifactId>spring-cloud-starter</artifactId>
+      <version>3.1.0</version>
+    </dependency>
+
 
     <dependency>
       <groupId>io.jaegertracing</groupId>
       <artifactId>jaeger-client</artifactId>
       <version>${jaeger.client.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>javax.inject</groupId>
+      <artifactId>javax.inject</artifactId>
+      <version>1</version>
     </dependency>
 
     <dependency>
@@ -153,7 +169,7 @@
     <plugins>
       <plugin>
         <artifactId>maven-war-plugin</artifactId>
-        <version>3.0.0</version>
+        <version>3.3.2</version>
         <configuration>
           <attachClasses>true</attachClasses>
           <classesClassifier>classes</classesClassifier>
