@@ -16,6 +16,7 @@
 
 package com.ntw.oms.admin.db;
 
+import com.datastax.oss.driver.api.core.CqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
@@ -23,10 +24,11 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
 import org.springframework.data.cassandra.config.*;
+import org.springframework.data.cassandra.core.CassandraOperations;
+import org.springframework.data.cassandra.core.CassandraTemplate;
 import org.springframework.data.cassandra.core.convert.CassandraConverter;
 import org.springframework.data.cassandra.core.convert.MappingCassandraConverter;
 import org.springframework.data.cassandra.core.cql.CqlTemplate;
-import org.springframework.data.cassandra.core.mapping.BasicCassandraMappingContext;
 import org.springframework.data.cassandra.core.mapping.CassandraMappingContext;
 import org.springframework.data.cassandra.core.mapping.SimpleUserTypeResolver;
 
@@ -46,45 +48,38 @@ public class CQLConfig {
 
     @Bean
     @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
-    public CassandraCqlClusterFactoryBean cluster() {
-        CassandraCqlClusterFactoryBean cluster = new CassandraCqlClusterFactoryBean();
-        cluster.setContactPoints(environment.getProperty("database.cassandra.hosts"));
-        cluster.setPort(Integer.parseInt(environment.getProperty("database.cassandra.port")));
-        return cluster;
-    }
-
-    /*
-     * Factory bean that creates the com.datastax.driver.core.Session instance
-     */
-    @Bean
-    @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
-    public CassandraCqlSessionFactoryBean session() {
-        CassandraCqlSessionFactoryBean session = new CassandraCqlSessionFactoryBean();
-        session.setCluster(cluster().getObject());
+    public CqlSessionFactoryBean session() {
+        CqlSessionFactoryBean session = new CqlSessionFactoryBean();
+        session.setContactPoints(environment.getProperty("database.cassandra.hosts"));
         session.setKeyspaceName(getKeyspaceName());
+        session.setLocalDatacenter("datacenter1");
         return session;
     }
 
     @Bean
     @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
-    public CqlTemplate cqlTemplate() {
-        return new CqlTemplate(session().getObject());
-    }
-
-    @Bean
-    @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
-    public CassandraMappingContext mappingContext() {
-        BasicCassandraMappingContext mappingContext =  new BasicCassandraMappingContext();
-        mappingContext.setUserTypeResolver(new SimpleUserTypeResolver(cluster().getObject(),
-                getKeyspaceName()));
-
+    public CassandraMappingContext mappingContext(CqlSession cqlSession) {
+        CassandraMappingContext mappingContext =  new CassandraMappingContext();
+        mappingContext.setUserTypeResolver(new SimpleUserTypeResolver(cqlSession));
         return mappingContext;
     }
 
     @Bean
     @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
-    public CassandraConverter converter() {
-        return new MappingCassandraConverter(mappingContext());
+    public CassandraConverter converter(CassandraMappingContext mappingContext) {
+        return new MappingCassandraConverter(mappingContext);
+    }
+
+    @Bean
+    @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
+    public CassandraOperations cassandraTemplate(CqlSession cqlSession) throws Exception {
+        return new CassandraTemplate(cqlSession);
+    }
+
+    @Bean
+    @ConditionalOnExpression("'${database.type}'.equals('CQL') or '${database.type}'.equals('ALL')")
+    public CqlTemplate cqlTemplate(CqlSession cqlSession) throws Exception {
+        return new CqlTemplate(cqlSession);
     }
 
 }
