@@ -16,6 +16,7 @@
 
 package com.ntw.oms.order;
 
+import com.datastax.oss.driver.api.core.CqlSession;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -24,18 +25,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
-import org.springframework.data.cassandra.config.CassandraClusterFactoryBean;
-import org.springframework.data.cassandra.config.CassandraCqlSessionFactoryBean;
-import org.springframework.data.cassandra.config.CassandraSessionFactoryBean;
-import org.springframework.data.cassandra.config.SchemaAction;
 import org.springframework.data.cassandra.core.CassandraOperations;
 import org.springframework.data.cassandra.core.CassandraTemplate;
-import org.springframework.data.cassandra.core.convert.CassandraConverter;
-import org.springframework.data.cassandra.core.convert.MappingCassandraConverter;
-import org.springframework.data.cassandra.core.cql.CqlTemplate;
-import org.springframework.data.cassandra.core.mapping.BasicCassandraMappingContext;
-import org.springframework.data.cassandra.core.mapping.CassandraMappingContext;
-import org.springframework.data.cassandra.core.mapping.SimpleUserTypeResolver;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import javax.sql.DataSource;
@@ -55,67 +46,11 @@ public class CartDBConfiguration {
     }
 
     @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraClusterFactoryBean cartDBCluster() {
-        CassandraClusterFactoryBean cluster = new CassandraClusterFactoryBean();
-        cluster.setContactPoints(environment.getProperty("database.cassandra.hosts"));
-        cluster.setPort(Integer.parseInt(environment.getProperty("database.cassandra.port")));
-        return cluster;
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraCqlSessionFactoryBean cartCqlSession() {
-        CassandraCqlSessionFactoryBean session = new CassandraCqlSessionFactoryBean();
-        session.setCluster(cartDBCluster().getObject());
-        session.setKeyspaceName(getKeyspaceName());
-        return session;
-    }
-
-    @Bean
-    @Qualifier("cartCqlTemplate")
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CqlTemplate cartCqlTemplate() throws Exception {
-        return new CqlTemplate(cartCqlSession().getObject());
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraMappingContext cartDBMappingContext() {
-        BasicCassandraMappingContext mappingContext =  new BasicCassandraMappingContext();
-        mappingContext.setUserTypeResolver(new SimpleUserTypeResolver(cartDBCluster().getObject(),
-                getKeyspaceName()));
-
-        return mappingContext;
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraConverter cartConverter() {
-        return new MappingCassandraConverter(cartDBMappingContext());
-    }
-
-    /*
-     * Factory bean that creates the com.datastax.driver.core.Session instance
-     */
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraSessionFactoryBean cartDBSession() throws Exception {
-        CassandraSessionFactoryBean session = new CassandraSessionFactoryBean();
-        session.setCluster(cartDBCluster().getObject());
-        session.setKeyspaceName(getKeyspaceName());
-        session.setConverter(cartConverter());
-        session.setSchemaAction(SchemaAction.NONE);
-        return session;
-    }
-
-    @Bean
     @Qualifier("cartCassandraOperations")
     @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraOperations cartCassandraTemplate() throws Exception {
-        return new CassandraTemplate(cartDBSession().getObject());
+    public CassandraOperations cartCassandraTemplate(CqlSession cqlSession) throws Exception {
+        return new CassandraTemplate(cqlSession);
     }
-
 
     private String getDriverClass() {
         return "org.postgresql.Driver";
