@@ -43,24 +43,12 @@ public class DBOrderLineDao implements OrderDao {
     @Qualifier("orderCassandraOperations")
     private CassandraOperations cassandraOperations;
 
-    @Autowired(required = false)
-    @Qualifier("orderCqlTemplate")
-    private CqlTemplate cqlTemplate;
-
     public CassandraOperations getCassandraOperations() {
         return cassandraOperations;
     }
 
     public void setCassandraOperations(CassandraOperations cassandraOperations) {
         this.cassandraOperations = cassandraOperations;
-    }
-
-    public CqlTemplate getCqlTemplate() {
-        return cqlTemplate;
-    }
-
-    public void setCqlTemplate(CqlTemplate cqlTemplate) {
-        this.cqlTemplate = cqlTemplate;
     }
 
     @Override
@@ -122,8 +110,7 @@ public class DBOrderLineDao implements OrderDao {
 
     @Override
     public boolean removeOrders() {
-        String cql = "truncate orderline";
-        cqlTemplate.execute(cql);
+        getCassandraOperations().truncate(DBOrderLine.class);
         logger.debug("Removed all orders");
         return true;
     }
