@@ -16,27 +16,26 @@
 
 package com.ntw.oms.product;
 
+import com.datastax.oss.driver.api.core.CqlSession;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
-import org.springframework.data.cassandra.config.CassandraClusterFactoryBean;
-import org.springframework.data.cassandra.config.CassandraCqlSessionFactoryBean;
-import org.springframework.data.cassandra.config.CassandraSessionFactoryBean;
+import org.springframework.data.cassandra.config.CqlSessionFactoryBean;
 import org.springframework.data.cassandra.config.SchemaAction;
+import org.springframework.data.cassandra.config.SessionFactoryFactoryBean;
 import org.springframework.data.cassandra.core.CassandraOperations;
 import org.springframework.data.cassandra.core.CassandraTemplate;
 import org.springframework.data.cassandra.core.convert.CassandraConverter;
 import org.springframework.data.cassandra.core.convert.MappingCassandraConverter;
-import org.springframework.data.cassandra.core.cql.CqlTemplate;
-import org.springframework.data.cassandra.core.mapping.BasicCassandraMappingContext;
 import org.springframework.data.cassandra.core.mapping.CassandraMappingContext;
 import org.springframework.data.cassandra.core.mapping.SimpleUserTypeResolver;
 import org.springframework.data.cassandra.repository.config.EnableCassandraRepositories;
+import org.springframework.data.mapping.context.MappingContext;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import javax.sql.DataSource;
@@ -45,7 +44,7 @@ import javax.sql.DataSource;
  * Created by anurag on 13/08/19.
  */
 @Configuration
-@EnableCassandraRepositories(basePackages = { "com.ntw.oms.product" })
+//@EnableCassandraRepositories(basePackages = { "com.ntw.oms.product" })
 @PropertySource(value = { "classpath:config.properties" })
 public class DBConfiguration {
 
@@ -58,62 +57,32 @@ public class DBConfiguration {
 
     @Bean
     @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraClusterFactoryBean cluster() {
-        CassandraClusterFactoryBean cluster = new CassandraClusterFactoryBean();
-        cluster.setContactPoints(environment.getProperty("database.cassandra.hosts"));
-        cluster.setPort(Integer.parseInt(environment.getProperty("database.cassandra.port")));
-        return cluster;
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraCqlSessionFactoryBean cqlSession() {
-        CassandraCqlSessionFactoryBean session = new CassandraCqlSessionFactoryBean();
-        session.setCluster(cluster().getObject());
+    public CqlSessionFactoryBean session() {
+        CqlSessionFactoryBean session = new CqlSessionFactoryBean();
+        session.setContactPoints(environment.getProperty("database.cassandra.hosts"));
         session.setKeyspaceName(getKeyspaceName());
+        session.setLocalDatacenter("datacenter1");
         return session;
     }
 
     @Bean
     @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CqlTemplate cqlTemplate() throws Exception {
-        return new CqlTemplate(cqlSession().getObject());
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraMappingContext mappingContext() {
-        BasicCassandraMappingContext mappingContext =  new BasicCassandraMappingContext();
-        mappingContext.setUserTypeResolver(new SimpleUserTypeResolver(cluster().getObject(),
-                getKeyspaceName()));
-
+    public CassandraMappingContext mappingContext(CqlSession cqlSession) {
+        CassandraMappingContext mappingContext =  new CassandraMappingContext();
+        mappingContext.setUserTypeResolver(new SimpleUserTypeResolver(cqlSession));
         return mappingContext;
     }
 
     @Bean
     @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraConverter converter() {
-        return new MappingCassandraConverter(mappingContext());
-    }
-
-    /*
-     * Factory bean that creates the com.datastax.driver.core.Session instance
-     */
-    @Bean
-    @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraSessionFactoryBean session() throws Exception {
-        CassandraSessionFactoryBean session = new CassandraSessionFactoryBean();
-        session.setCluster(cluster().getObject());
-        session.setKeyspaceName(getKeyspaceName());
-        session.setConverter(converter());
-        session.setSchemaAction(SchemaAction.NONE);
-        return session;
+    public CassandraConverter converter(CassandraMappingContext mappingContext) {
+        return new MappingCassandraConverter(mappingContext);
     }
 
     @Bean
     @ConditionalOnProperty(name = "database.type", havingValue = "CQL")
-    public CassandraOperations cassandraTemplate() throws Exception {
-        return new CassandraTemplate(session().getObject());
+    public CassandraOperations cassandraTemplate(CqlSession cqlSession) throws Exception {
+        return new CassandraTemplate(cqlSession);
     }
 
     private String getDriverClass() {
