@@ -22,7 +22,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.cassandra.core.CassandraOperations;
-import org.springframework.data.cassandra.core.cql.CqlTemplate;
 import org.springframework.stereotype.Component;
 
 import java.util.LinkedList;
@@ -39,23 +38,12 @@ public class DBProductDao implements ProductDao {
     @Autowired(required = false)
     private CassandraOperations cassandraOperations;
 
-    @Autowired(required = false)
-    private CqlTemplate cqlTemplate;
-
     public CassandraOperations getCassandraOperations() {
         return cassandraOperations;
     }
 
     public void setCassandraOperations(CassandraOperations cassandraOperations) {
         this.cassandraOperations = cassandraOperations;
-    }
-
-    public CqlTemplate getCqlTemplate() {
-        return cqlTemplate;
-    }
-
-    public void setCqlTemplate(CqlTemplate cqlTemplate) {
-        this.cqlTemplate = cqlTemplate;
     }
 
     @Override
@@ -70,7 +58,7 @@ public class DBProductDao implements ProductDao {
         }
         if (products.isEmpty()) {
             logger.error("Unable to get Products; context={}", allProductCql);
-            return null;
+            return new LinkedList<>();
         }
         logger.debug("Fetched all products; context={}", products);
         return products;
@@ -156,7 +144,7 @@ public class DBProductDao implements ProductDao {
     public boolean removeProducts() {
         String removeProductsCql = "truncate product";
         try {
-            getCassandraOperations().delete(removeProductsCql);
+            getCassandraOperations().truncate(Product.class);
         } catch (Exception e) {
             logger.error("Exception deleting products: ", e);
             return false;
